renamed file to InventoryCellRenderer
